debug

+++ src/main/java/com/westlake/air/pecs/domain/bean/score/FeatureByPep.java
@@ -17,7 +17,7 @@ public class FeatureByPep {
 
     List<List<ExperimentFeature>> experimentFeatures;
 
-    List<Float> libraryIntensityList;
+    List<Double> libraryIntensityList;
 
     List<double[]> noise1000List;
 }
